Revert "Revert "static_cast directly from double to enum""
This reapplies 842a18440197b3163e284ab5e21ed7f8c0d0b85b before merging
the changes that depend on it in master.

+++ gpt_input.cpp
@@ -404,9 +404,7 @@ void gpt_input::DoHarmonize()
 
     oenum_alb_or_anb const alb_anb =
         static_cast<oenum_alb_or_anb>
-            (static_cast<int>
-                (database_->Query(DB_AgeLastOrNearest)
-                )
+            (database_->Query(DB_AgeLastOrNearest)
             );
     DateOfBirth.minimum_and_maximum
         (minimum_birthdate(IssueAge.maximum(), EffectiveDate.value(), alb_anb)
@@ -550,9 +548,7 @@ void gpt_input::DoTransmogrify()
 
     oenum_alb_or_anb const alb_anb =
         static_cast<oenum_alb_or_anb>
-            (static_cast<int>
-                (database_->Query(DB_AgeLastOrNearest)
-                )
+            (database_->Query(DB_AgeLastOrNearest)
             );
 
     int apparent_age = attained_age
